Move remaining common models
