+++ src/backend/TrafficCourts/Citizen.Service/TrafficCourts/Common/OpenAPIs/OracleDataApi/v1_0/OcrViolationTicket.cs
@@ -1,0 +1,6 @@
+﻿namespace TrafficCourts.Common.OpenAPIs.OracleDataApi.v1_0
+{
+    internal class OcrViolationTicket
+    {
+    }
+}
