+++ src/backend/TrafficCourts/Citizen.Service/Validators/FormRecognizerValidator.cs
@@ -1,8 +1,9 @@
 using TrafficCourts.Citizen.Service.Validators.Rules;
 using TrafficCourts.Common.Features.Lookups;
-using TrafficCourts.Common.OpenAPIs.OracleDataApi.v1_0;
 using System.Text.RegularExpressions;
-using ViolationTicketVersion = TrafficCourts.Domain.Models.ViolationTicketVersion;
+using TrafficCourts.Domain.Models;
+
+using OcrViolationTicket = TrafficCourts.Domain.Models.OcrViolationTicket;
 
 namespace TrafficCourts.Citizen.Service.Validators;
