+++ src/backend/TrafficCourts/Citizen.Service/Validators/Rules/CountSectionRule.cs
@@ -1,6 +1,6 @@
 using System.Text.RegularExpressions;
 using TrafficCourts.Common.Features.Lookups;
-using TrafficCourts.Common.OpenAPIs.OracleDataApi.v1_0;
+using TrafficCourts.Domain.Models;
 
 namespace TrafficCourts.Citizen.Service.Validators.Rules;
