+++ src/backend/TrafficCourts/Citizen.Service/Validators/Rules/DateOfServiceLT30Rule.cs
@@ -1,4 +1,4 @@
-using TrafficCourts.Common.OpenAPIs.OracleDataApi.v1_0;
+using TrafficCourts.Domain.Models;
 
 namespace TrafficCourts.Citizen.Service.Validators.Rules;
