+++ src/backend/TrafficCourts/Citizen.Service/Validators/Rules/LowConfidenceGlobalRule.cs
@@ -1,6 +1,5 @@
-using TrafficCourts.Common.OpenAPIs.OracleDataApi.v1_0;
-using static TrafficCourts.Common.OpenAPIs.OracleDataApi.v1_0.OcrViolationTicket;
-using ViolationTicketVersion = TrafficCourts.Domain.Models.ViolationTicketVersion;
+using TrafficCourts.Domain.Models;
+using static TrafficCourts.Domain.Models.OcrViolationTicket;
 
 namespace TrafficCourts.Citizen.Service.Validators.Rules;
