+++ src/backend/TrafficCourts/Test/Citizen.Service/Validators/Rules/CountActRegMustBeMVATest.cs
@@ -4,7 +4,7 @@ using System.Threading.Tasks;
 using TrafficCourts.Citizen.Service.Validators;
 using TrafficCourts.Citizen.Service.Validators.Rules;
 using TrafficCourts.Common.Features.Lookups;
-using TrafficCourts.Common.OpenAPIs.OracleDataApi.v1_0;
+using TrafficCourts.Domain.Models;
 using Xunit;
 
 namespace TrafficCourts.Test.Citizen.Service.Validators.Rules;
