+++ src/backend/TrafficCourts/Test/Citizen.Service/Validators/Rules/FieldMatchesRegexRuleTest.cs
@@ -1,6 +1,6 @@
 using System.Threading.Tasks;
 using TrafficCourts.Citizen.Service.Validators.Rules;
-using TrafficCourts.Common.OpenAPIs.OracleDataApi.v1_0;
+using TrafficCourts.Domain.Models;
 using Xunit;
 
 namespace TrafficCourts.Test.Citizen.Service.Validators.Rules;
